Create UploadAttachmentsWorker to be wrapped in android worker
